requirements.txt
opencv-python
numpy
Pillow
streamlit
# System-level dependencies
libgl1-mesa-glx